Created view for CompanyNameDetails
seperated out the CompanyNameDetails functionality into it's own file within src/components/views. Imported it into CobdForm.vue which will be the main page which changes views (steps in wizard)
